AOS-6591 refactoring to separate functionality with and without nexus access token
